# docs/reference/SMART_AO_V8_ASSIGNMENT_OPENAPI.json
{
  "components": {
    "schemas": {
...
      "AcknowledgeAssignmentRequest": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "format": "uuid",
            "title": "Command Id",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "command_id",
          "idempotency_key",
          "expected_revision"
        ],
        "title": "AcknowledgeAssignmentRequest",
        "type": "object"
      },
      "AggregateReferenceResponse": {
        "additionalProperties": false,
        "properties": {
          "aggregate_id": {
            "format": "uuid",
            "title": "Aggregate Id",
            "type": "string"
          },
          "aggregate_revision": {
            "minimum": 0.0,
            "title": "Aggregate Revision",
            "type": "integer"
          },
          "aggregate_type": {
            "title": "Aggregate Type",
            "type": "string"
          }
        },
        "required": [
          "aggregate_type",
          "aggregate_id",
          "aggregate_revision"
        ],
        "title": "AggregateReferenceResponse",
        "type": "object"
      },
...
      "AssignmentCommandResponse": {
        "additionalProperties": false,
        "properties": {
          "aggregate_refs": {
            "items": {
              "$ref": "#/components/schemas/AggregateReferenceResponse"
            },
            "title": "Aggregate Refs",
            "type": "array"
          },
          "command_id": {
            "format": "uuid",
            "title": "Command Id",
            "type": "string"
          },
          "event_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          },
          "result_code": {
            "enum": [
              "ASSIGNMENT_ACKNOWLEDGED",
              "ASSIGNMENT_CLARIFICATION_REQUESTED",
              "ASSIGNMENT_UNAVAILABILITY_REPORTED"
            ],
            "title": "Result Code",
            "type": "string"
          },
          "status": {
            "const": "SUCCEEDED",
            "default": "SUCCEEDED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "idempotency_key",
          "result_code",
          "aggregate_refs",
          "event_ids"
        ],
        "title": "AssignmentCommandResponse",
        "type": "object"
      },
...
      "ReportAssignmentUnavailabilityRequest": {
        "additionalProperties": false,
        "properties": {
          "command_id": {
            "format": "uuid",
            "title": "Command Id",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "impact_note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Note"
          },
          "known_deadline_impact": {
            "default": false,
            "title": "Known Deadline Impact",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "reason_kind": {
            "enum": [
              "SICKNESS",
              "LEAVE",
              "CAPACITY_CONFLICT",
              "SKILL_GAP",
              "ACCESS_PROBLEM",
              "OTHER"
            ],
            "title": "Reason Kind",
            "type": "string"
          },
          "unavailable_from": {
            "format": "date-time",
            "title": "Unavailable From",
            "type": "string"
          },
          "unavailable_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Until"
          }
        },
        "required": [
          "command_id",
          "idempotency_key",
          "expected_revision",
          "reason_kind",
          "reason",
          "unavailable_from"
        ],
        "title": "ReportAssignmentUnavailabilityRequest",
        "type": "object"
      },
      "RequestAssignmentClarificationRequest": {
        "additionalProperties": false,
        "properties": {
          "clarification_kind": {
            "enum": [
              "SCOPE",
              "PRIORITY",
              "DEADLINE",
              "DOCUMENT",
              "RESPONSIBILITY",
              "OTHER"
            ],
            "title": "Clarification Kind",
            "type": "string"
          },
          "command_id": {
            "format": "uuid",
            "title": "Command Id",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "format": "uuid",
            "title": "Idempotency Key",
            "type": "string"
          },
          "priority": {
            "default": "NORMAL",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH"
            ],
            "title": "Priority",
            "type": "string"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "requested_scope": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Scope"
          },
          "subject": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "idempotency_key",
          "expected_revision",
          "clarification_kind",
          "subject",
          "question"
        ],
        "title": "RequestAssignmentClarificationRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SaaS BTP d'analyse DCE et de décision d'appel d'offres.",
    "title": "SMART_AO V8",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assignments/{assignment_id}/acknowledgement": {
      "post": {
        "operationId": "acknowledge_assignment_api_v1_assignments__assignment_id__acknowledgement_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcknowledgeAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Rejeu idempotent de la commande déjà appliquée."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Bearer absent, invalide ou expiré."
          },
          "403": {
            "description": "Capability ou scope ReBAC insuffisant."
          },
          "404": {
            "description": "Affectation absente ou hors tenant : réponse volontairement neutre."
          },
          "409": {
            "description": "Clé d’idempotence réutilisée ou commande encore en cours."
          },
          "422": {
            "description": "Validation de requête ou rejet métier de la commande."
          }
        },
        "summary": "Acknowledge Assignment",
        "tags": [
          "assignments"
        ]
      }
    },
    "/api/v1/assignments/{assignment_id}/clarification-requests": {
      "post": {
        "operationId": "request_clarification_api_v1_assignments__assignment_id__clarification_requests_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestAssignmentClarificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Rejeu idempotent de la commande déjà appliquée."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Bearer absent, invalide ou expiré."
          },
          "403": {
            "description": "Capability ou scope ReBAC insuffisant."
          },
          "404": {
            "description": "Affectation absente ou hors tenant : réponse volontairement neutre."
          },
          "409": {
            "description": "Clé d’idempotence réutilisée ou commande encore en cours."
          },
          "422": {
            "description": "Validation de requête ou rejet métier de la commande."
          }
        },
        "summary": "Request Clarification",
        "tags": [
          "assignments"
        ]
      }
    },
...
    "/api/v1/assignments/{assignment_id}/unavailability-reports": {
      "post": {
        "operationId": "report_unavailability_api_v1_assignments__assignment_id__unavailability_reports_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportAssignmentUnavailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Rejeu idempotent de la commande déjà appliquée."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Bearer absent, invalide ou expiré."
          },
          "403": {
            "description": "Capability ou scope ReBAC insuffisant."
          },
          "404": {
            "description": "Affectation absente ou hors tenant : réponse volontairement neutre."
          },
          "409": {
            "description": "Clé d’idempotence réutilisée ou commande encore en cours."
          },
          "422": {
            "description": "Validation de requête ou rejet métier de la commande."
          }
        },
        "summary": "Report Unavailability",
        "tags": [
          "assignments"
        ]
      }
    }
  }
}